+++ test/spec/collection/sphinx.rb
@@ -1,0 +1,3 @@
+shared_examples 'collection::sphinx' do
+    include_examples 'sphinx::sphinx'
+end