+++ test/spec/shared/base/packages.rb
@@ -1,17 +1,7 @@
 shared_examples 'base::packages' do
     [
-        'openssh-server',
-        'dnsmasq',
-        'postfix',
-        'unzip',
-        'bzip2',
         'wget',
         'curl',
-        'moreutils',
-        'rsync',
-        'git',
-        'nano',
-        'net-tools',
     ].each do |package|
         describe package("#{package}") do
             it { should be_installed }