+++ test/spec/shared/bootstrap/layout.rb
@@ -36,6 +36,8 @@ shared_examples 'bootstrap::layout' do
     ## Scripts
     #########################
     [
+        "/usr/local/bin/apk-install",
+        "/usr/local/bin/apk-upgrade",
         "/usr/local/bin/apt-install",
         "/usr/local/bin/apt-upgrade",
         "/usr/local/bin/service",