+++ test/spec/shared/bootstrap/toolchain.rb
@@ -4,7 +4,11 @@ shared_examples 'bootstrap::toolchain' do
         its(:exit_status) { should eq 0 }
     end
 
-    describe command('easy_install --version') do
+    describe command('easy_install --version'), :if => ['debian', 'ubuntu', 'redhat'].include?(os[:family]) do
+        its(:exit_status) { should eq 0 }
+    end
+
+    describe command('easy_install-2.7 --version'), :if => ['alpine'].include?(os[:family]) do
         its(:exit_status) { should eq 0 }
     end